further code refactor for application construction to allowed for customisable styling
