+++ apps.scss
@@ -24,7 +24,7 @@
     top: 0;
     width: 100%;
     height: 5%;
-    background-color: darkslategrey;
+    background-color: blueviolet;
 
     &__controls {
         text-align: center;
