+++ styles.css
@@ -190,7 +190,7 @@
   top: 0;
   width: 100%;
   height: 5%;
-  background-color: darkslategrey;
+  background-color: blueviolet;
 }
 .app-head__controls {
   text-align: center;
@@ -242,6 +242,14 @@
   display: flex;
   justify-content: center;
   align-items: center;
+}
+
+.notes-app-container {
+  border: 1px solid #2b2a2a;
+}
+.notes-app-container__head {
+  border-bottom: 1px solid lightgray;
+  background-color: white;
 }
 
 body {
